+++ apps/web-v2/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @meeshy/web-v2
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @meeshy/shared@1.21.3
+
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @meeshy/shared@1.21.2
+
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @meeshy/shared@1.21.1
+
 ## 2.0.6
 
 ### Patch Changes